+++ frontend/src/router/index.ts
@@ -788,21 +788,21 @@ const routes: Array<RouteRecordRaw> = [
         path: "",
         name: "sql-editor.home",
         meta: { title: () => "SQL Editor" },
-        component: () => import("../views/SqlEditor/SqlEditor.vue"),
+        component: () => import("../views/sql-editor/SqlEditor.vue"),
         props: true,
       },
       {
         path: "/sql-editor/:connectionSlug",
         name: "sql-editor.detail",
         meta: { title: () => "SQL Editor" },
-        component: () => import("../views/SqlEditor/SqlEditor.vue"),
+        component: () => import("../views/sql-editor/SqlEditor.vue"),
         props: true,
       },
       {
         path: "/sql-editor/:connectionSlug/:sheetSlug",
         name: "sql-editor.share",
         meta: { title: () => "SQL Editor" },
-        component: () => import("../views/SqlEditor/SqlEditor.vue"),
+        component: () => import("../views/sql-editor/SqlEditor.vue"),
         props: true,
       },
     ],
