+++ frontend/src/views/sql-editor/AsidePanel/AsidePanel.vue
