+++ frontend/src/views/sql-editor/AsidePanel/QueryHistoryContainer.vue
