+++ frontend/src/views/sql-editor/EditorHeader.vue
