+++ frontend/src/views/sql-editor/EditorPanel/ConnectionHolder.vue
