+++ frontend/src/views/sql-editor/EditorPanel/EditorAction.vue
