+++ frontend/src/views/sql-editor/EditorPanel/EditorPanel.vue
