+++ frontend/src/views/sql-editor/EditorPanel/QueryEditor.vue
