+++ frontend/src/views/sql-editor/EditorPanel/SharePopover.vue
