+++ frontend/src/views/sql-editor/GettingStarted.vue
