+++ frontend/src/views/sql-editor/TabListContainer.vue
